+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to the "xssencode" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.3] - 2021/04/04
+
+### Added
+
+- String => Bash
+- String => PowerShell
+- String => Python
+- String => Perl
 
 ## [0.0.2] - 2020/12/11